Create Airplane Status Indicator Application
In this commit, a simple application has been developed using an airplane as an example. The application utilizes user LEDs to visually represent the status of the airplane, providing a visual indicator for various conditions and states.

The key features of this application include:

If the aeroplane rotates more than 90 degree on either left or right the user led1 should high
If the aeroplane rotates more than 50 degree on either forward or backward the user led2 should high
If the aeroplane near to ground the user led3 should high
If the temerature in aeroplane is higher than 27c all 3 user